+++ src/external_services/google_sheets.py
@@ -2,9 +2,9 @@ import pygsheets
 import pandas as pd
 
 from src.constants.robinhood_constants import (
-    DEFAULT_SPREADSHEET_NAME,
-    SheetHeaders
+    DEFAULT_SPREADSHEET_NAME
 )
+from src.constants.report_constants import SheetHeaders
 
 
 def initialize_column_headers(worksheet: pygsheets.Worksheet) -> None:
